DefaultReadFile properly implements the IStreamIterator interface
Fixes #138
