+++ news/138.bugfix
@@ -1,1 +1,2 @@
 Use mock from unittest on Python 3 [ale-rt]
+DefaultReadFile properly implements the IStreamIterator interface [ale-rt]
